Widen int to size_t in pixel loops

+++ src/iop/monochrome.c
@@ -212,7 +212,7 @@ void process(dt_iop_module_t *self,
   const float d_a = d->a;
   const float d_b = d->b;
   DT_OMP_FOR_SIMD(aligned(in, out:64))
-  for(int k = 0; k < 4*npixels; k += 4)
+  for(size_t k = 0; k < 4*npixels; k += 4)
   {
     out[k+0] = 100.0f * _color_filter(in[k+1], in[k+2], d_a, d_b, sigma2);
     out[k+1] = out[k+2] = 0.0f;
@@ -243,7 +243,7 @@ void process(dt_iop_module_t *self,
 
   const float highlights = d->highlights;
   DT_OMP_FOR_SIMD(aligned(in, out:64))
-  for(int k = 0; k < 4*npixels; k += 4)
+  for(size_t k = 0; k < 4*npixels; k += 4)
   {
     const float tt = _envelope(in[k]);
     const float t = tt + (1.0f - tt) * (1.0f - highlights);
